+++ go/chat/teams.go
@@ -117,6 +117,7 @@ func (t *ImplicitTeamsNameInfoSource) Lookup(ctx context.Context, name string, v
 	if strings.HasPrefix(name, keybase1.ImplicitTeamPrefix) {
 		return t.lookupInternalName(ctx, name, vis)
 	}
+	res = types.NewNameInfo()
 
 	// Always create here to simulate behavior of GetTLFCryptKeys
 	teamID, _, impTeamName, _, err := teams.LookupOrCreateImplicitTeam(ctx, t.G().ExternalG(), name,
@@ -128,7 +129,6 @@ func (t *ImplicitTeamsNameInfoSource) Lookup(ctx context.Context, name string, v
 		panic(fmt.Sprintf("implicit team found via LookupImplicitTeam not root team: %s", teamID))
 	}
 
-	res = types.NewNameInfo()
 	res.CanonicalName = impTeamName.String()
 	res.ID, err = chat1.TeamIDToTLFID(teamID)
 	if err != nil {
@@ -206,4 +206,53 @@ func (t *ImplicitTeamsNameInfoSource) lookupInternalName(ctx context.Context, na
 	}
 
 	return res, nil
+}
+
+func LoadTeam(ctx context.Context, g *libkb.GlobalContext, tlfID chat1.TLFID,
+	membersType chat1.ConversationMembersType, public bool,
+	loadTeamArgOverride func(keybase1.TeamID) keybase1.LoadTeamArg) (team *teams.Team, err error) {
+
+	// Set up load team argument construction, possibly controlled by the caller
+	ltarg := func(teamID keybase1.TeamID) keybase1.LoadTeamArg {
+		return keybase1.LoadTeamArg{
+			ID:     teamID,
+			Public: public,
+		}
+	}
+	if loadTeamArgOverride != nil {
+		ltarg = loadTeamArgOverride
+	}
+
+	switch membersType {
+	case chat1.ConversationMembersType_IMPTEAMNATIVE, chat1.ConversationMembersType_TEAM:
+		teamID, err := keybase1.TeamIDFromString(tlfID.String())
+		if err != nil {
+			return team, err
+		}
+		return teams.Load(ctx, g, ltarg(teamID))
+	case chat1.ConversationMembersType_IMPTEAMUPGRADE:
+		arg := libkb.NewAPIArgWithNetContext(ctx, "team/id")
+		arg.Args.Add("tlf_id", libkb.S{Val: tlfID.String()})
+		res, err := g.API.Get(arg)
+		if err != nil {
+			return team, err
+		}
+		st, err := res.Body.AtKey("team_id").GetString()
+		if err != nil {
+			return team, err
+		}
+		teamID, err := keybase1.TeamIDFromString(st)
+		if err != nil {
+			return team, err
+		}
+		team, err = teams.Load(ctx, g, ltarg(teamID))
+		if err != nil {
+			return team, err
+		}
+		if !tlfID.EqString(team.KBFSTLFID()) {
+			return team, fmt.Errorf("mismatch TLFID to team: %s != %s", team.KBFSTLFID(), tlfID)
+		}
+		return team, nil
+	}
+	return team, fmt.Errorf("invalid impteam members type: %v", membersType)
 }
